api-server/requirements.txt
Flask
Flask-SQLAlchemy
PyMuPDF 
langchain
faiss-cpu
google-generativeai
tiktoken
werkzeug
langchain-community
flask_cors
langchain_google_genai
numpy
